feat: add shared neo4j and io utilities
